fix: push result of builtin op to stash
Currently, results of builtin functions are not pushed to the
`Stash`, leading to bugs in programs with multiple `CallExpression`
statements. Since we pop from the `Stash` after every
`ExpressionStatement`, by not pushing results of builtin function
calls to the `Stash`, we will be incorrectly popping from the
`Stash` after every builtin `CallExpression` statement.