feat: parse time UID for AST nodes (#14)
* feat: add `uid` to AST nodes at parse time

Add the `--cache` flag is necessary to ensure that there are no gaps in
the uid sequence. I am not entirely why disabling cache creates gaps,
but it seems like its due to parser retries?

* feat: add AST node location at parse time

* refactor: include `uid` and `loc` in ECE `Node` interface

Set both as optional fields for now to not deal with errors. `loc` field
type definition does not include `source` sub field since we are not
dealing with multiple files