feat: implement support for `ClosureOp` in heap
Note that the newly added `AstMap` to the program context is lazy,
meaning only when a reference is needed, will we keep track of it.
Additionally, I have also remove `FuncDeclOp` as I found it to be a
redundant indirection. The new refactor to `Environment` also allows us
to remove the need to return environments from the interpreter since it
is all in-place ops now.