feat: integrate `Stash` with heap (#15)
* chore: add docs for `isAny` util func

* refactor: `Stack` push and pop via class methods

* fix: explicitly check for `undefined` instead of bool check

Right now, I am using a boolean check for `undefined` but this also
means that a valid `false` value will be treated as a nil value.

* refactor: restructure `Heap`

Place configuration params and tags into separate files for better
separation of concern. Added `alloc` and `resolve` as proof of concept
generic methods to allocate tagged pointers in the heap and translating
heap address to underlying data/value respectively

Added support for `Number` tagged pointer for testing. Also added more
docs to describe `Heap` methods.

* feat: integrate `Stash` with `Heap`

Note that not all values that can possibly appear in the `Stash` can be
allocated in the heap yet! Right now, we are just adding the extra steps
of allocating to the heap and resolving heap address to `Stash`
operations

* feat: add support for booleans in heap

* refactor: implement `tag` helper function to tag of buffer block

* feat: implement support for `BuiltinOp` in heap

* feat: allocate reusable `bool` literal value at start of heap

This will bring down memory usage if we were to do many relational
checks

* refactor: rename `name` field to `id` for `FunctionDeclaration` node

* refactor: introduce the construct of `Frame` to `Environment`

This give us the ability to identify a frame with a numeric id instead
of just using an object reference. As such, we will be able to serialize
a reference to a environment in the heap

* feat: implement support for `ClosureOp` in heap

Note that the newly added `AstMap` to the program context is lazy,
meaning only when a reference is needed, will we keep track of it.
Additionally, I have also remove `FuncDeclOp` as I found it to be a
redundant indirection. The new refactor to `Environment` also allows us
to remove the need to return environments from the interpreter since it
is all in-place ops now.

* chore: missing newline