refactor: abstract `AstMap` as a class object
This is mainly so that we are able to maintain a state to keep track of
what is the next `uid` for a AST node. We need to produce more uid since
we may do runtime transpilation (e.g. `ForClause` -> `ForCondition`) that
introduces more AST nodes. A more straightforward example would be the
inject of `main()` call expression to the AST before the execution,
where the `CallExpression` needs a uid.