feat: integrate `Control` with heap (#18)
* refactor: abstract `AstMap` as a class object

This is mainly so that we are able to maintain a state to keep track of
what is the next `uid` for a AST node. We need to produce more uid since
we may do runtime transpilation (e.g. `ForClause` -> `ForCondition`) that
introduces more AST nodes. A more straightforward example would be the
inject of `main()` call expression to the AST before the execution,
where the `CallExpression` needs a uid.

* chore: remove `husky` for better DX 🚀

`pre-commit` and `pre-push` hooks are super slow and we don't have any
tests right now so I decided to remove it temporarily (we can always
revert it later).

* feat: implement `AstNode` and `CallOp` support in the heap

Added helper functions in the heap to allocate/resolve multiple values.
This would prevent having to `H.(alloc/resolve).bind(H)` multiple times
in the ECE

* feat: integrate AST nodes and `CallOp` with heap

Bunch of minor refactors to make code more readable

* fix: update uid of `CALL_MAIN` for each evaluate

Currently, we are doing a in-place uid assign to `CALL_MAIN` and we only
assign a uid if there no uid in the node. However, since `CALL_MAIN` in
the outer scope of `evaluate`, we do not assign the right uid on each
run of `evaluate`, since the `SourceFile` uid changes.

* feat: implement support for `EnvOp` in heap

* feat: pass all `Control` values through heap

* feat: implement support for `VarDeclOp` in heap

Added some helper functions to `AstMap`

* feat: implement `UnaryOp` and `BinaryOp` support in heap

* feat: implement `AssignOp` support in heap

* feat: implement `PopSOp` support in heap

* refactor: remove redundant indirection for `BuiltinOp`

Instead of having a extra `ApplyBuiltinOp`, we execute the op during the
handling of `CallOp`

* refactor: add type assertion for clarity