feat: implement `GoStatement` support in ECE
Right now we only support non-builtin calls to spawn a goroutine. We
need to consider if its even worth it spawn a new goroutine for a
builtin since its only a "1 step/tick" op. Additionally, I introduced
`GoRoutineOp` to handle spawning of new goroutines. Its payload is
same as `CallOp` so we can consider combining them if there is no
special cases to handle later on