feat: goroutines (#19)
* feat: naive time quanta based scheduler

Abstracted the current program as a `main` go routine in preparation for
`GoStatement` implementation

* chore: fix eslint errors

* refactor: use `Counter` for frame id generation

* refactor: add `GoStatement` to ECE types

* feat: implement `GoStatement` support in ECE

Right now we only support non-builtin calls to spawn a goroutine. We
need to consider if its even worth it spawn a new goroutine for a
builtin since its only a "1 step/tick" op. Additionally, I introduced
`GoRoutineOp` to handle spawning of new goroutines. Its payload is
same as `CallOp` so we can consider combining them if there is no
special cases to handle later on

* feat: add `StringLit` support in parser

This is mainly so that we can `println` strings to differentiate between
goroutines. Note that this is purely a convenience feature, the ECE does
not allocate strings in the heap/string pool right now.

* feat: implement `GoRoutineOp` support in heap

* feat: randomize goroutine time quanta